Let a discard record a fractional quantity (#860)
* Let a discard record a fractional quantity

A half-installed conduit coil had no representable exit. The shelf could
hold `0.5 each` because `InventoryEntry.amount` has always been a free
float, but `Expense.productQuantity` was an `integer` column with
`z.number().int()` in front of it and `step="1"` on the field — so the
smallest discard was 1, and every partially-used product sat in "Shelf
disagrees" forever because the variance could never reach zero.

Widen the column to `double precision`, matching `cost` and every other
measure in the schema. The CHECK constraint only tests `<> 0` and
nullability, so it survives unchanged — no hand-applied ALTER.

The column type alone changes nothing: the `::int` casts in
quantity-ledger, pricing and related-view are the real boundary, and
would have truncated 0.5 back to 0 at the SELECT. `pricing.ts` matters
most — that cast is the denominator of the derived unit price, which
flows into `InventoryEntry.valuation`.

`unitWeighted` needed the `::numeric` cast `costWeighted` already
carried. `unitWeight` is `numeric`, and Postgres has no
`double precision * numeric` operator, so without it the Product list
fails at runtime rather than at typecheck.

`Product.expectedQuantity` and `ProductComponent.quantity` stay integral
— the first is a one-of-a-kind flag tested only as `= 1`, the second is
kit cardinality.

Also warn on the inventory checkbox, which silently soft-deleted a whole
shelf entry: when the discard empties the shelf, when it exceeds what the
shelf holds, and when declining the checkbox leaves a row claiming stock
just written off. Over-discard stays allowed rather than refused the way
bulk.ts refuses an over-move — Tenet 1 makes the shelf a stale-tolerant
ballpark, and nothing server-side can tell a stale count from a fat
finger. The reasoning is recorded at the branch. Its `remaining < 0` path
was untested and indistinguishable from `remaining === 0`; it is pinned
now.

The success toast read `Discarded 3 × Widget` whether or not the shelf
row was destroyed. `inventory.removed` was dead payload; it now says so.


* Let the expense quantity filter bounds be fractional too

`Test auxiliary packages` caught what `pnpm check` and the web vitest
projects do not run: packages/schemas has its own suite, and one case
still asserted the old whole-number contract.

`productQuantityMin` of 1.5 moves from the rejection list to a new
acceptance case. A bound that can only be whole cannot bracket a
fractional quantity at all, so the filter would silently exclude the very
rows it exists to find — and `productQuantityMax: -1` is the "everything
written off" worklist, which a fractional discard now lands in.

The `it.each` bundled two unrelated claims. Split rather than deleted:
"not-a-number" is still rejected, and the string case is kept as its own
acceptance to cover the URL/MCP coercion path.


---------

+++ packages/schemas/src/problems.ts
@@ -118,9 +118,9 @@ export const productMissingPriceSchema = z.object({
 export const negativeExpectedQuantitySchema = z.object({
   ...productProblemFields,
   /** Negative by construction — that is the defect. */
-  expectedQuantity: z.number().int(),
-  acquiredUnits: z.number().int(),
-  exitedUnits: z.number().int(),
+  expectedQuantity: z.number(),
+  acquiredUnits: z.number(),
+  exitedUnits: z.number(),
   unknownAcquisitionLines: z.number().int(),
   unknownExitLines: z.number().int(),
 });
